Data.Tree

Description

Multi-way trees (aka rose trees) and forests.

Documentation

data Tree a

Multi-way trees, also known as rose trees.

Constructors

Node

Fields

rootLabel :: a

label value

subForest :: Forest a

zero or more child trees

type Forest a = [Tree a]

drawTree :: Tree String -> String

Neat 2-dimensional drawing of a tree.

draw :: Tree String -> [String]

flatten :: Tree a -> [a]

The elements of a tree in pre-order.

flattenPostOrder :: Tree a -> [a]

The elements of the tree in post-order

subTrees :: Tree a -> [Tree a]

List of subtrees (including the tree itself), in pre-order.

subTrees = flatten . cojoin

For post-order:

subtreesPostOrder = flattenPostOrder . cojoin

levels :: Tree a -> [[a]]

Lists of nodes at each level of the tree.

subTreesByLevel :: Tree a -> [[Tree a]]

List of subtrees at each level of the tree.

subTreesByLevel = levels . cojoin

unfoldTree :: (b -> (a, [b])) -> b -> Tree a

Build a tree from a seed value

unfoldTreeM :: Monad m => (b -> m (a, [b])) -> b -> m (Tree a)

Monadic tree builder, in depth-first order

unfoldTreeM_BF :: Monad m => (b -> m (a, [b])) -> b -> m (Tree a)

Monadic tree builder, in breadth-first order, using an algorithm adapted from Breadth-First Numbering: Lessons from a Small Exercise in Algorithm Design, by Chris Okasaki, ICFP'00.

lookupTree :: Eq a => a -> Tree a -> Maybe (Tree a)

Get the sub-tree rooted at the first (left-most, depth-first) occurrence of the specified node value

For other traversal strategies, consider the following definition

lookupTree i = (Data.List.find ((==i) . rootLabel) . flatten . cojoin

lookupTreeBy :: (a -> Bool) -> Tree a -> Maybe (Tree a)

Get the sub-tree rooted at the first (left-most, depth-first) node that fulfils the given condition.

For other traversal strategies, consider the following definition

lookupTreeBy f = (Data.List.find (f . rootLabel) . flatten . cojoin

filterPruneTree :: (a -> Bool) -> Tree a -> Maybe (Tree a)

Prune every subtree whose root label does not match.

filterGraftTree :: (a -> Bool) -> Tree a -> Forest a

Remove nodes that do not match, and graft the children of the removed node onto the tree in place of the parent.

size :: Tree a -> Int

Number of nodes on the three

maxDepth :: Tree a -> Int

Maximum depth of tree

prune :: Int -> Tree a -> Tree a

remove all nodes past a certain depth

mirror :: Tree a -> Tree a

Reverse every forest within the three

cojoin :: Tree a -> Tree (Tree a)

Label each node of the tree with its full subtree.

This is the duplicate function from the Tree Comonad.